backend/foodgram/requirements.txt
Django==3.2.16
requests==2.26.0
Pillow==8.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
djoser==2.1.0
drf-extra-fields==3.7.0
python-dotenv==0.20.0
django-cors-headers==3.13.0
django-filter==23.3
psycopg2-binary==2.9.3
gunicorn==20.1.0
flake8==7.0.0